requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
openai
python-multipart